+++ packages/analytics/analytics-geo-map/CHANGELOG.md
@@ -3,6 +3,48 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.9.15](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Fanalytics-geo-map%401.9.14...%40kong-ui-public%2Fanalytics-geo-map%401.9.15) (2026-08-22)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-geo-map
+
+
+
+
+
+## [1.9.14](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Fanalytics-geo-map%401.9.13...%40kong-ui-public%2Fanalytics-geo-map%401.9.14) (2026-08-22)
+
+### Bug Fixes
+
+* **deps:** update non-major [@kong](https://github.com/kong) dependencies with stable versions ([#3723](https://github.com/Kong/public-ui-components/issues/3723)) ([df7e955](https://github.com/Kong/public-ui-components/commit/df7e95536ba2f2ea8d46fa1de541b163b5b1c126))
+
+
+## [1.9.13](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Fanalytics-geo-map%401.9.12...%40kong-ui-public%2Fanalytics-geo-map%401.9.13) (2026-08-21)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-geo-map
+
+
+
+
+
+## [1.9.12](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-geo-map@1.9.11...@kong-ui-public/analytics-geo-map@1.9.12) (2026-08-20)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-geo-map
+
+
+
+
+
+## [1.9.11](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-geo-map@1.9.10...@kong-ui-public/analytics-geo-map@1.9.11) (2026-08-19)
+
+
+### Bug Fixes
+
+* **analytics:** respect app theme in map colors [MA-5263] ([#3709](https://github.com/Kong/public-ui-components/issues/3709)) ([6e143df](https://github.com/Kong/public-ui-components/commit/6e143df5cb95babca205aa31ab1fa2e4312348a1))
+
+
+
+
+
 ## [1.9.10](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-geo-map@1.9.9...@kong-ui-public/analytics-geo-map@1.9.10) (2026-08-18)
 
 **Note:** Version bump only for package @kong-ui-public/analytics-geo-map