+++ packages/core/table-data-grid/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.13](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Ftable-data-grid%400.3.12...%40kong-ui-public%2Ftable-data-grid%400.3.13) (2026-08-22)
+
+### Bug Fixes
+
+* **deps:** update all non-major dependencies with stable versions ([#3731](https://github.com/Kong/public-ui-components/issues/3731)) ([79e2e2d](https://github.com/Kong/public-ui-components/commit/79e2e2d53fdd6362c899464ee44a27bdf5370a14))
+
+
+## [0.3.12](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Ftable-data-grid%400.3.11...%40kong-ui-public%2Ftable-data-grid%400.3.12) (2026-08-22)
+
+### Bug Fixes
+
+* **deps:** update non-major [@kong](https://github.com/kong) dependencies with stable versions ([#3723](https://github.com/Kong/public-ui-components/issues/3723)) ([df7e955](https://github.com/Kong/public-ui-components/commit/df7e95536ba2f2ea8d46fa1de541b163b5b1c126))
+
+
 ## [0.3.11](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/table-data-grid@0.3.10...@kong-ui-public/table-data-grid@0.3.11) (2026-08-14)
 
 **Note:** Version bump only for package @kong-ui-public/table-data-grid