+++ packages/entities/entities-certificates/CHANGELOG.md
@@ -3,6 +3,60 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.16.34](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Fentities-certificates%403.16.33...%40kong-ui-public%2Fentities-certificates%403.16.34) (2026-08-24)
+
+**Note:** Version bump only for package @kong-ui-public/entities-certificates
+
+
+
+
+
+## [3.16.33](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Fentities-certificates%403.16.32...%40kong-ui-public%2Fentities-certificates%403.16.33) (2026-08-24)
+
+**Note:** Version bump only for package @kong-ui-public/entities-certificates
+
+
+
+
+
+## [3.16.32](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Fentities-certificates%403.16.31...%40kong-ui-public%2Fentities-certificates%403.16.32) (2026-08-22)
+
+**Note:** Version bump only for package @kong-ui-public/entities-certificates
+
+
+
+
+
+## [3.16.31](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Fentities-certificates%403.16.30...%40kong-ui-public%2Fentities-certificates%403.16.31) (2026-08-22)
+
+### Bug Fixes
+
+* **deps:** update all non-major dependencies with stable versions ([#3731](https://github.com/Kong/public-ui-components/issues/3731)) ([79e2e2d](https://github.com/Kong/public-ui-components/commit/79e2e2d53fdd6362c899464ee44a27bdf5370a14))
+
+
+## [3.16.30](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Fentities-certificates%403.16.29...%40kong-ui-public%2Fentities-certificates%403.16.30) (2026-08-22)
+
+**Note:** Version bump only for package @kong-ui-public/entities-certificates
+
+
+
+
+
+## [3.16.29](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Fentities-certificates%403.16.28...%40kong-ui-public%2Fentities-certificates%403.16.29) (2026-08-22)
+
+### Bug Fixes
+
+* **deps:** update non-major [@kong](https://github.com/kong) dependencies with stable versions ([#3723](https://github.com/Kong/public-ui-components/issues/3723)) ([df7e955](https://github.com/Kong/public-ui-components/commit/df7e95536ba2f2ea8d46fa1de541b163b5b1c126))
+
+
+## [3.16.28](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-certificates@3.16.27...@kong-ui-public/entities-certificates@3.16.28) (2026-08-21)
+
+**Note:** Version bump only for package @kong-ui-public/entities-certificates
+
+
+
+
+
 ## [3.16.27](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-certificates@3.16.26...@kong-ui-public/entities-certificates@3.16.27) (2026-08-14)
 
 **Note:** Version bump only for package @kong-ui-public/entities-certificates